requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
et-xmlfile==1.1.0
Flask==2.2.2
frozenlist==1.3.3
future==0.18.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
liffpy==0.0.1a3
line-bot-sdk==2.3.0
MarkupSafe==2.1.1
multidict==6.0.3
mypy-extensions==0.4.3
numpy==1.24.0
openai==0.25.0
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.221213
pathspec==0.10.3
platformdirs==2.6.0
pymongo==4.3.3
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
six==1.16.0
tomli==2.0.1
tqdm==4.64.1
types-pytz==2022.7.0.0
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
yarl==1.8.2
